requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-keycloak==3.7.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0
